chore: Address review change requests
